+++ frontend/src/metabase/lib/email.ts
@@ -1,7 +1,10 @@
+// Enhanced email regex with Unicode support for international characters
+// Uses Unicode character classes to support international domains and names
+// Based on RFC 5322 with practical constraints for security and usability
 const EMAIL_REGEX =
-  /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  /^(?=.{1,254}$)(?=.{1,64}@)[\p{L}\p{N}!#$%&'*+\/=?\^`{|}~_-]+(?:\.[\p{L}\p{N}!#$%&'*+\/=?\^`{|}~_-]+)*@[\p{L}\p{N}](?:[\p{L}\p{N}-]{0,61}[\p{L}\p{N}])?(?:\.[\p{L}\p{N}](?:[\p{L}\p{N}-]{0,61}[\p{L}\p{N}])?)*$/u;
 
-export function isEmail(email: string | undefined | null) {
+export function isEmail(email: string | undefined | null): boolean {
   if (email === null || email === undefined) {
     return false;
   }
