+++ frontend/src/metabase/lib/formatting/email.tsx
@@ -1,15 +1,10 @@
 import ExternalLink from "metabase/common/components/ExternalLink";
+import { isEmail } from "metabase/lib/email";
 import { getDataFromClicked } from "metabase-lib/v1/parameters/utils/click-behavior";
 
 import { renderLinkTextForClick } from "./link";
 import { removeNewLines } from "./strings";
 import type { OptionsType } from "./types";
-
-// Enhanced email regex with Unicode support for international characters
-// Uses Unicode character classes to support international domains and names
-// Based on RFC 5322 with practical constraints for security and usability
-const EMAIL_ALLOW_LIST_REGEX =
-  /^(?=.{1,254}$)(?=.{1,64}@)[\p{L}\p{N}!#$%&'*+\/=?\^`{|}~_-]+(?:\.[\p{L}\p{N}!#$%&'*+\/=?\^`{|}~_-]+)*@[\p{L}\p{N}](?:[\p{L}\p{N}-]{0,61}[\p{L}\p{N}])?(?:\.[\p{L}\p{N}](?:[\p{L}\p{N}-]{0,61}[\p{L}\p{N}])?)*$/u;
 
 export function formatEmail(
   value: string,
@@ -32,7 +27,7 @@ export function formatEmail(
     jsx &&
     rich &&
     (view_as === "email_link" || view_as === "auto") &&
-    EMAIL_ALLOW_LIST_REGEX.test(email)
+    isEmail(email)
   ) {
     let displayText = label || email;
     if (collapseNewlines) {
