+++ frontend/src/metabase/lib/formatting/email.unit.spec.tsx
@@ -163,86 +163,5 @@ describe("formatEmail", () => {
       expect(formatEmail(null as any)).toBe("null");
       expect(formatEmail(undefined as any)).toBe("undefined");
     });
-
-    it("should work with email addresses containing ASCII special characters", () => {
-      const specialEmails = [
-        "user+tag@example.com",
-        "user.name@example.com",
-        "user_name@example.com",
-        "user-name@example.com",
-        "user123@example.com",
-      ];
-
-      specialEmails.forEach((email) => {
-        const result = formatEmail(email, { jsx: true, rich: true });
-        expect(isElementOfType(result as JSX.Element, ExternalLink)).toBe(true);
-
-        render(result);
-
-        expect(screen.getByRole("link")).toHaveAttribute(
-          "href",
-          `mailto:${email}`,
-        );
-
-        cleanup();
-      });
-    });
-
-    it("should handle emails with Unicode/international characters", () => {
-      const unicodeEmails = [
-        "hafthór_júlíus_björnsson@gameofthron.es", // Original example with accented characters
-        "test.用户@example.com", // Chinese characters
-        "josé.garcía@café.org", // Spanish accented characters
-        "müller@münchen.de", // German umlauts
-        "andré@québec.ca", // French accented characters
-      ];
-
-      unicodeEmails.forEach((email) => {
-        const result = formatEmail(email, { jsx: true, rich: true });
-        expect(isElementOfType(result as JSX.Element, ExternalLink)).toBe(true);
-        render(result);
-        expect(screen.getByRole("link")).toHaveAttribute(
-          "href",
-          `mailto:${email}`,
-        );
-        expect(screen.getByRole("link")).toHaveTextContent(email);
-        cleanup();
-      });
-    });
-
-    it("should handle emails with quoted local parts", () => {
-      // These are technically valid emails according to RFC 5322 but may not pass the regex
-      const quotedEmails = [
-        '"user name"@example.com',
-        '"special.chars"@example.com',
-        '"test@test"@example.com',
-      ];
-
-      quotedEmails.forEach((email) => {
-        const result = formatEmail(email, { jsx: true, rich: true });
-        // Document current behavior - these likely won't pass the current regex
-        expect(result).toBe(email);
-      });
-    });
-
-    it("should handle emails with unusual but valid characters", () => {
-      const unusualEmails = [
-        "user!#$%&'*+-/=?^_`{|}~@example.com", // All allowed special chars in local part
-        "x@example-domain.com",
-        "test@sub.domain.example.org",
-      ];
-
-      unusualEmails.forEach((email) => {
-        const result = formatEmail(email, { jsx: true, rich: true });
-
-        expect(isElementOfType(result as JSX.Element, ExternalLink)).toBe(true);
-        render(result);
-        expect(screen.getByRole("link")).toHaveAttribute(
-          "href",
-          `mailto:${email}`,
-        );
-        cleanup();
-      });
-    });
   });
 });
